added balance bar and new events

+++ app/src/main/java/com/example/explory/data/model/quest/DistanceQuestDto.kt
@@ -2,5 +2,5 @@ package com.example.explory.data.model.quest
 
 data class DistanceQuestDto(
     val commonQuestDto: QuestDto,
-    val distance: Long
+    val distance: Double
 )
