+++ app/src/main/java/com/example/explory/data/model/quest/RouteDto.kt
@@ -1,5 +1,5 @@
 package com.example.explory.data.model.quest
 
 data class RouteDto(
-    val points: List<PointDto>, val distance: Long, val routeId: Long
+    val points: List<PointDto>, val distance: Double, val routeId: Long
 )
