+++ app/src/main/java/com/example/explory/presentation/screen/quest/DistanceQuestScreen.kt
@@ -26,7 +26,7 @@ fun DistanceQuestScreen(
     description: String,
     difficulty: String,
     transportType: String,
-    distance: Long,
+    distance: Double,
     image: String?,
     onNavigateBack: () -> Unit
 ) {
